tests/requirements.txt
# E2E Testing Requirements

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-xdist==3.3.1  # Parallel test execution
pytest-html==4.1.1   # HTML reports
pytest-json-report==1.5.0

# Kubernetes client
kubernetes==28.1.0
kubernetes-asyncio==28.2.0

# HTTP clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Data and validation
pydantic==2.5.2
jsonschema==4.20.0
pyyaml==6.0.1

# Test data generation
faker==20.1.0
factory-boy==3.3.0

# Metrics and monitoring
prometheus-client==0.19.0
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
tenacity==8.2.3  # Retry logic

# Load testing (optional, for Python-based tests)
locust==2.17.0